feat: Address PR comments
