+++ integration_tests/mqtt/docker-compose.yml
@@ -6,7 +6,11 @@ services:
       file: ../../docker/docker-compose.yml
       service: risingwave-standalone
   mqtt-server:
-    image: emqx/emqx:5.2.1
+    image: eclipse-mosquitto
+    command:
+    - sh
+    - -c
+    - echo "running command"; printf 'allow_anonymous true\nlistener 1883 0.0.0.0' > /mosquitto/config/mosquitto.conf; echo "starting service..."; cat /mosquitto/config/mosquitto.conf;/docker-entrypoint.sh;/usr/sbin/mosquitto -c /mosquitto/config/mosquitto.conf
     ports:
       - 1883:1883
   etcd-0:
