+++ src/connector/src/lib.rs
@@ -52,6 +52,7 @@ pub mod sink;
 pub mod source;
 
 pub mod common;
+pub mod mqtt_common;
 
 pub use paste::paste;
